Use separate node for SDF passo
Not strictly necessary, but I do like having an explicit render graph,
especially if it's something someone else ever wanted to tweak.
