+++ src/render/lighting/node.rs
@@ -11,17 +11,10 @@ use bevy::render::renderer::RenderDevice;
 use bevy::render::view::{ViewTarget, ViewUniformOffset, ViewUniforms};
 use smallvec::{smallvec, SmallVec};
 
-use crate::render::extract::{
-    ExtractedAmbientLight2d, ExtractedLightOccluder2d, ExtractedPointLight2d,
-};
+use crate::render::extract::{ExtractedAmbientLight2d, ExtractedPointLight2d};
 
-use super::{
-    LightMapPipeline, Lighting2dAuxiliaryTextures, LightingPipeline, LightingPipelineId,
-    SdfPipeline,
-};
+use super::{LightMapPipeline, Lighting2dAuxiliaryTextures, LightingPipeline, LightingPipelineId};
 
-const SDF_PASS: &str = "sdf_pass";
-const SDF_BIND_GROUP: &str = "sdf_bind_group";
 const LIGHT_MAP_PASS: &str = "light_map_pass";
 const LIGHT_MAP_BIND_GROUP: &str = "light_map_bind_group";
 const LIGHTING_PASS: &str = "lighting_pass";
@@ -49,21 +42,17 @@ impl ViewNode for LightingNode {
         >,
         world: &'w World,
     ) -> Result<(), bevy::render::render_graph::NodeRunError> {
-        let sdf_pipeline_resource = world.resource::<SdfPipeline>();
         let light_map_pipeline_resource = world.resource::<LightMapPipeline>();
         let pipeline = world.resource::<LightingPipeline>();
         let pipeline_cache = world.resource::<PipelineCache>();
 
         let (
-            Some(sdf_pipeline),
             Some(lighting_pipeline),
             Some(light_map_pipeline),
             Some(view_uniform_binding),
             Some(ambient_light_uniform),
             Some(point_light_binding),
-            Some(light_occluders_binding),
         ) = (
-            pipeline_cache.get_render_pipeline(sdf_pipeline_resource.pipeline_id),
             pipeline_cache.get_render_pipeline(pipeline_id.0),
             pipeline_cache.get_render_pipeline(light_map_pipeline_resource.pipeline_id),
             world.resource::<ViewUniforms>().uniforms.binding(),
@@ -74,51 +63,10 @@ impl ViewNode for LightingNode {
             world
                 .resource::<GpuArrayBuffer<ExtractedPointLight2d>>()
                 .binding(),
-            world
-                .resource::<GpuArrayBuffer<ExtractedLightOccluder2d>>()
-                .binding(),
         )
         else {
             return Ok(());
         };
-
-        // SDF
-        let bind_group = render_context.render_device().create_bind_group(
-            SDF_BIND_GROUP,
-            &sdf_pipeline_resource.layout,
-            &BindGroupEntries::sequential((view_uniform_binding.clone(), light_occluders_binding)),
-        );
-
-        let mut sdf_pass = render_context.begin_tracked_render_pass(RenderPassDescriptor {
-            label: Some(SDF_PASS),
-            color_attachments: &[Some(RenderPassColorAttachment {
-                view: &aux_textures.sdf.default_view,
-                resolve_target: None,
-                ops: Operations::default(),
-            })],
-            ..default()
-        });
-
-        let mut dynamic_offsets: SmallVec<[u32; 3]> = smallvec![view_offset.offset];
-
-        // Storage buffers aren't available in WebGL2. We fall back to a
-        // dynamic uniform buffer, and therefore need to provide the offset.
-        // We're providing a value of 0 here as we're limiting the number of
-        // point lights to only those that can reasonably fit in a single binding.
-        if world
-            .resource::<RenderDevice>()
-            .limits()
-            .max_storage_buffers_per_shader_stage
-            == 0
-        {
-            dynamic_offsets.push(0);
-        }
-
-        sdf_pass.set_render_pipeline(sdf_pipeline);
-        sdf_pass.set_bind_group(0, &bind_group, &dynamic_offsets);
-        sdf_pass.draw(0..3, 0..1);
-
-        drop(sdf_pass);
 
         // Light map
         let light_map_bind_group = render_context.render_device().create_bind_group(
