+++ src/render/lighting/pipeline.rs
@@ -11,67 +11,14 @@ use bevy::render::renderer::RenderDevice;
 use bevy::render::texture::BevyDefault;
 use bevy::render::view::{ViewTarget, ViewUniform};
 
-use crate::render::extract::{
-    ExtractedAmbientLight2d, ExtractedLightOccluder2d, ExtractedPointLight2d,
-};
+use crate::render::extract::{ExtractedAmbientLight2d, ExtractedPointLight2d};
 
-use super::{LightingPipelineKey, LIGHTING_SHADER, LIGHT_MAP_SHADER, SDF_SHADER};
+use super::{LightingPipelineKey, LIGHTING_SHADER, LIGHT_MAP_SHADER};
 
-const SDF_PIPELINE: &str = "sdf_pipeline";
-const SDF_BIND_GROUP_LAYOUT: &str = "sdf_bind_group_layout";
 const LIGHTING_PIPELINE: &str = "lighting_pipeline";
 const LIGHTING_BIND_GROUP_LAYOUT: &str = "lighting_bind_group_layout";
 const LIGHT_MAP_BIND_GROUP_LAYOUT: &str = "light_map_group_layout";
 const LIGHT_MAP_PIPELINE: &str = "light_map_pipeline";
-
-#[derive(Resource)]
-pub struct SdfPipeline {
-    pub layout: BindGroupLayout,
-    pub pipeline_id: CachedRenderPipelineId,
-}
-
-impl FromWorld for SdfPipeline {
-    fn from_world(world: &mut World) -> Self {
-        let render_device = world.resource::<RenderDevice>();
-        let pipeline_cache = world.resource::<PipelineCache>();
-
-        let layout = render_device.create_bind_group_layout(
-            SDF_BIND_GROUP_LAYOUT,
-            &BindGroupLayoutEntries::sequential(
-                ShaderStages::FRAGMENT,
-                (
-                    uniform_buffer::<ViewUniform>(true),
-                    GpuArrayBuffer::<ExtractedLightOccluder2d>::binding_layout(render_device),
-                ),
-            ),
-        );
-
-        let pipeline_id = pipeline_cache.queue_render_pipeline(RenderPipelineDescriptor {
-            label: Some(SDF_PIPELINE.into()),
-            layout: vec![layout.clone()],
-            vertex: fullscreen_shader_vertex_state(),
-            fragment: Some(FragmentState {
-                shader: SDF_SHADER,
-                shader_defs: vec![],
-                entry_point: "fragment".into(),
-                targets: vec![Some(ColorTargetState {
-                    format: TextureFormat::Rgba16Float,
-                    blend: None,
-                    write_mask: ColorWrites::ALL,
-                })],
-            }),
-            primitive: PrimitiveState::default(),
-            depth_stencil: None,
-            multisample: MultisampleState::default(),
-            push_constant_ranges: vec![],
-        });
-
-        Self {
-            layout,
-            pipeline_id,
-        }
-    }
-}
 
 #[derive(Resource)]
 pub struct LightMapPipeline {
