+++ src/render/mod.rs
@@ -1,2 +1,3 @@
 pub mod extract;
 pub mod lighting;
+pub mod sdf;
